cleanup: move functions only used in tests to test files
This commit moves the firstGraphemeCluster() and
firstGraphemeClusterInString() functions to test files (since that is
where they are used).
